Fixed compiler warning.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@6104 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/race/race_manager.cpp
@@ -212,7 +212,7 @@ void RaceManager::startNew()
     m_kart_status.clear();
 
     // Before thi
-    assert(m_num_karts == m_ai_kart_list.size()+m_player_karts.size());
+    assert((unsigned int)m_num_karts == m_ai_kart_list.size()+m_player_karts.size());
     // First add the AI karts (randomly chosen)
     // ----------------------------------------
     int init_gp_rank = 0;
